Defer loop count loading.
Calling CGImageSourceCopyProperties is very slow on larger GIFs (consuming ~400 milliseconds on iPhone 5 with 1100 frame 240x320 GIF). Loop count isn't needed until the first run of a GIF has ended, so don't load it eagerly. If the GIF ends before the loop count has loaded it pauses at the final frame until it is.